requirements.txt
attrs==19.1.0
python-rocksdb==0.7.0
Whoosh==2.7.4
click==7.0